consolidate diverged examples
commit #1
for now there is one class that defines the examples
and two classes providing the backend

for the moment only hdf5 works
need to figure out a good way to do xml as well (or alternatively)

I'm not committing the generated files. Should they be part of the repository?